requirements.txt
redis==4.5.4
grpcio==1.56.0
grpcio-tools==1.37.1
pony==0.7.16
pymysql==1.1.0
retry
openai
gradio==3.36.1
langchain~=0.1.11
pyyaml~=6.0.1
aioredis==2.0.1